+++ src/app/base/pages/educatio-evidences-list/educatio-evidences-list.component.ts
@@ -4,6 +4,8 @@ import { WorkShopsFilter } from "../../models";
 import { ModalService } from "../../../shared/services";
 import { AgGridInterFace } from "../../../shared/interfaces/ag-grid.interface";
 import { propertyOf } from "../../../shared/utilities/property-of";
+import { ConfirmInterFace } from "../../../shared/ki-components/ki-confirmation/confirm.interface";
+
 import {
   EducationEvidencesFormModalComponent,
   TaxFormModalComponent,
@@ -99,13 +101,22 @@ export class EducationEvidencesListComponent implements OnInit {
   }
 
   removeCell() {
-    if (this.selectRow.length) {
-      for (let i = 0; i <= this.selectRow.length; i++) {
-        this.onDeleteItem(this.selectRow[i]);
+    const params: ConfirmInterFace = {
+      acceptText: "بله",
+      declineText: "خیر",
+      description: "آیا از عملیات مورد نظر اطمینان دارید؟",
+      title: "حذف" + " " + `"${this.selectRow[0].evidence.toUpperCase()}"`,
+      type: "Confirm",
+    };
+    this._modalService.showConfirm(params, false).then((res) => {
+      if (res) {
+        if (this.selectRow.length) {
+          for (let i = 0; i <= this.selectRow.length; i++) {
+            this.onDeleteItem(this.selectRow[i]);
+          }
+        }
       }
-    } else {
-      //  this._toaster.error("لطفا یک رکورد انتخاب شود");
-    }
+    });
   }
 
   onDeleteItem(item: EducationEvidencesDto) {
