+++ src/app/base/pages/tax-list/tax-list.component.ts
@@ -10,6 +10,7 @@ import { TaxFormModalComponent } from "../../components/templates";
 import { ChangeWorkShopsService } from "../../../services/change-work-shop.service";
 import { finalize } from "rxjs";
 import { TaxDto } from "../../models/tax.model";
+import { ConfirmInterFace } from "../../../shared/ki-components/ki-confirmation/confirm.interface";
 
 @Component({
   selector: "app-tax-list",
@@ -102,14 +103,24 @@ export class TaxListComponent implements OnInit {
       error: (err) => {},
     });
   }
+
   removeCell() {
-    if (this.selectRow.length) {
-      for (let i = 0; i <= this.selectRow.length; i++) {
-        this.onDeleteItem(this.selectRow[i]);
+    const params: ConfirmInterFace = {
+      acceptText: "بله",
+      declineText: "خیر",
+      description: "آیا از عملیات مورد نظر اطمینان دارید؟",
+      title: "حذف" + " " + `"${this.selectRow[0].name.toUpperCase()}"`,
+      type: "Confirm",
+    };
+    this._modalService.showConfirm(params, false).then((res) => {
+      if (res) {
+        if (this.selectRow.length) {
+          for (let i = 0; i <= this.selectRow.length; i++) {
+            this.onDeleteItem(this.selectRow[i]);
+          }
+        }
       }
-    } else {
-      //  this._toaster.error("لطفا یک رکورد انتخاب شود");
-    }
+    });
   }
   onDeleteItem(item: BenefitDeductionDto) {
     this.isShowLoadingDelete = true;
